refactor(project): Restructure modules into packages with __init__.py
